+++ packages/snaps-simulation/src/store/index.ts
@@ -5,3 +5,4 @@ export * from './state';
 export * from './store';
 export * from './trackables';
 export * from './ui';
+export * from './web-sockets';